Split PopularComment into its own component
