+++ packages/lesswrong/lib/components.ts
@@ -545,6 +545,7 @@ importComponent("SingleLineComment", () => require('../components/comments/Singl
 importComponent("ShowParentComment", () => require('../components/comments/ShowParentComment'));
 importComponent("NewUserGuidelinesDialog", () => require('../components/comments/NewUserGuidelinesDialog'));
 importComponent("PopularCommentsList", () => require('../components/comments/PopularCommentsList'));
+importComponent("PopularComment", () => require('../components/comments/PopularComment'));
 
 importComponent("PostsListEditorSearchHit", () => require('../components/search/PostsListEditorSearchHit'));
 importComponent("PostsSearchHit", () => require('../components/search/PostsSearchHit'));
